unvalid bug fixed

+++ app/src/main/java/com/teamducky/ecocentric/HomeActivity.java
@@ -64,7 +64,7 @@ public class HomeActivity extends AppCompatActivity implements EasyPermissions.P
                 startService(intent);
             }
         }
-        allPointsTxt.setText((int)ParseUser.getCurrentUser().get("points")+"");
+        allPointsTxt.setText(ParseUser.getCurrentUser().get("points")+"");
     }
 
     @Override
